refactoring for tests
